requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
future==0.18.2
gql==2.0.0
graphql-core==2.3.2
idna==2.10
promise==2.3
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==12.4.2
requests==2.25.1
Rx==1.6.1
six==1.16.0
tornado==6.1
urllib3==1.26.5
